requirements.txt
flet==0.80.5
pg8000==1.31.5
python-dotenv==1.2.1
